requirement.txt
librosa
praat-parselmouth
numpy
pandas
torch
torchaudio
scikit-learn